cleanup
renames: 
- onPurchaseTicketV3 => onPurchaseTicket
 -startTicketBuyerV3Attempt => startTicketBuyerAttempts
- STARTTICKETBUYERV3_* => STARTTICKETBUYER_
- purchaseTicketsV3 => purchaseTickets
- onV3PurchaseTicket => onPurchaseTicket
- startTicketAutoBuyerV3 => startTicketAutoBuyer

clean purchaseTab
PurchaseTicketsForm, PurchaseTicketsForm is not a shared component any more is not a shared component any more
