--- app/components/shared/TicketAutoBuyerForm/index.js
@@ -1,1 +1,0 @@
-export { default } from "./TicketAutoBuyerForm";
