+++ app/components/shared/index.js
@@ -22,8 +22,6 @@ export { default as PrivacyForm } from "./PrivacyForm/PrivacyForm";
 export { default as Log } from "./Log/Log";
 export { default as SendTransaction } from "./SendTransaction/SendTransaction";
 export { default as Collapse } from "./Collapse";
-export { default as TicketAutoBuyerForm } from "./TicketAutoBuyerForm";
-export { default as PurchaseTicketsForm } from "./PurchaseTicketsForm";
 export { default as AnimatedContainer } from "./AnimatedContainer";
 export { default as LNInvoiceStatus } from "./LNInvoiceStatus";
 export { default as LNPaymentStatus } from "./LNPaymentStatus";
