+++ app/components/views/TicketsPage/PurchaseTab/PurchaseTab.jsx
@@ -1,7 +1,148 @@
-import PurchasePage from "./PurchaseTickets/PurchaseTickets";
+import { useEffect, useState } from "react";
+import { PurchaseTabPage } from "./PurchaseTabPage";
+import { usePurchaseTab } from "./hooks";
+import { isEqual } from "lodash/fp";
 
-function Purchase() {
-  return <PurchasePage />;
-}
+const PurchaseTab = () => {
+  const {
+    spvMode,
+    blocksNumberToNextTicket,
+    sidebarOnBottom,
+    isWatchingOnly,
+    // TODO add retry wiith discoverAvailableVsps in case of failure:
+    // discoverAvailableVSPs,
+    availableVSPs,
+    // TODO treat errors:
+    // availableVSPsError,
+    defaultSpendingAccount,
+    ticketPrice,
+    purchaseTicketsAttempt,
+    mixedAccount,
+    changeAccount,
+    isLoading,
+    rememberedVspHost,
+    setRememberedVspHost,
+    notMixedAccounts,
+    isVSPListingEnabled,
+    onEnableVSPListing,
+    getRunningIndicator,
+    visibleAccounts
+  } = usePurchaseTab();
 
-export default Purchase;
+  const [account, setAccount] = useState(defaultSpendingAccount);
+  useEffect(() => {
+    const newAccount = visibleAccounts?.find((a) =>
+      isEqual(a.value, account?.value)
+    );
+    newAccount && setAccount(newAccount);
+  }, [visibleAccounts, account]);
+
+  // todo use this vsp to buy solo tickets.
+  const [vsp, setVSP] = useState(() => {
+    if (rememberedVspHost) {
+      // reset rememberedVspHost if it's outdated
+      if (
+        availableVSPs?.find(
+          (availableVSP) => availableVSP.host === rememberedVspHost.host
+        )?.outdated === true
+      ) {
+        setRememberedVspHost(null);
+        return null;
+      } else {
+        return { host: rememberedVspHost.host };
+      }
+    } else {
+      return null;
+    }
+  });
+  const [numTicketsToBuy, setNumTicketsToBuy] = useState(1);
+  const [isValid, setIsValid] = useState(false);
+
+  const toggleRememberVspHostCheckBox = () => {
+    setRememberedVspHost(!rememberedVspHost ? vsp : null);
+  };
+
+  const handleOnKeyDown = (e) => {
+    if (e.keyCode == 38) {
+      e.preventDefault();
+      onIncrementNumTickets();
+    } else if (e.keyCode == 40) {
+      e.preventDefault();
+      onDecrementNumTickets();
+    }
+  };
+
+  const onIncrementNumTickets = () => {
+    setNumTicketsToBuy((numTicketsToBuy) =>
+      numTicketsToBuy == "" ? 1 : numTicketsToBuy + 1
+    );
+  };
+
+  const onChangeNumTickets = (numTicketsToBuy) => {
+    if (parseInt(numTicketsToBuy)) {
+      setNumTicketsToBuy(parseInt(numTicketsToBuy));
+    } else if (numTicketsToBuy == "") {
+      setNumTicketsToBuy(numTicketsToBuy);
+    }
+  };
+
+  const onDecrementNumTickets = () => {
+    setNumTicketsToBuy((numTicketsToBuy) =>
+      numTicketsToBuy <= 1 ? 1 : numTicketsToBuy - 1
+    );
+  };
+
+  const onPurchaseTicket = (passphrase) => {
+    purchaseTicketsAttempt(passphrase, account, numTicketsToBuy, vsp);
+  };
+
+  useEffect(() => {
+    const { spendable } = account;
+    const canAfford = numTicketsToBuy * ticketPrice <= spendable;
+    const hasTickets = numTicketsToBuy > 0;
+    setIsValid(canAfford && hasTickets && !!vsp);
+  }, [ticketPrice, numTicketsToBuy, account, vsp]);
+
+  const [vspFee, setVspFee] = useState(
+    availableVSPs &&
+      availableVSPs.find((availableVSP) => availableVSP.host === vsp?.host)
+        ?.vspData.feepercentage
+  );
+
+  return (
+    <PurchaseTabPage
+      {...{
+        spvMode,
+        blocksNumberToNextTicket,
+        sidebarOnBottom,
+        isWatchingOnly,
+        account,
+        numTicketsToBuy,
+        onIncrementNumTickets,
+        onDecrementNumTickets,
+        onChangeNumTickets,
+        handleOnKeyDown,
+        setAccount,
+        ticketPrice,
+        isValid,
+        availableVSPs,
+        setVSP,
+        onPurchaseTicket,
+        vsp,
+        vspFee,
+        setVspFee,
+        mixedAccount,
+        changeAccount,
+        isLoading,
+        rememberedVspHost,
+        toggleRememberVspHostCheckBox,
+        notMixedAccounts,
+        isVSPListingEnabled,
+        onEnableVSPListing,
+        getRunningIndicator
+      }}
+    />
+  );
+};
+
+export default PurchaseTab;
