+++ app/components/views/TicketsPage/PurchaseTab/TicketAutoBuyer/TicketAutoBuyer.module.css
