+++ app/components/views/TicketsPage/PurchaseTab/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PurchaseTab";
