+++ app/reducers/vsp.js
@@ -39,9 +39,9 @@ import {
   SET_NUM_TICKETS_TO_BUY
 } from "actions/VSPActions";
 import {
-  STARTTICKETBUYERV3_ATTEMPT,
-  STARTTICKETBUYERV3_SUCCESS,
-  STARTTICKETBUYERV3_FAILED,
+  STARTTICKETBUYER_ATTEMPT,
+  STARTTICKETBUYER_SUCCESS,
+  STARTTICKETBUYER_FAILED,
   STOPTICKETBUYER_SUCCESS
 } from "actions/ControlActions";
 import { CLOSEWALLET_SUCCESS } from "actions/WalletLoaderActions";
@@ -60,9 +60,9 @@ export default function vsp(state = {}, action) {
         availableVSPs: null,
         availableVSPsError: action.error
       };
-    case STARTTICKETBUYERV3_ATTEMPT:
+    case STARTTICKETBUYER_ATTEMPT:
       return { ...state, ticketAutoBuyerRunning: false };
-    case STARTTICKETBUYERV3_SUCCESS:
+    case STARTTICKETBUYER_SUCCESS:
       return {
         ...state,
         ticketAutoBuyerRunning: true,
@@ -71,7 +71,7 @@ export default function vsp(state = {}, action) {
         balanceToMaintain: action.balanceToMaintain,
         account: action.account.name
       };
-    case STARTTICKETBUYERV3_FAILED:
+    case STARTTICKETBUYER_FAILED:
       return { ...state, ticketAutoBuyerRunning: false };
     case STOPTICKETBUYER_SUCCESS:
       return { ...state, ticketAutoBuyerRunning: false };
