+++ test/unit/components/views/TicketsPage/PurchaseTab/TicketsPage.spec.js
@@ -29,7 +29,6 @@ let mockWalletCfgGet;
 let mockWalletCfgSet;
 
 beforeEach(() => {
-  selectors.stakePoolListingEnabled = jest.fn(() => true);
   selectors.getAvailableVSPs = jest.fn(() => mockAvailableVsps);
   selectors.spendingAccounts = jest.fn(() => [mockMixedAccount]);
   selectors.visibleAccounts = jest.fn(() => [mockMixedAccount]);
